+++ login.php
@@ -9,6 +9,12 @@ require_once __DIR__.'/init.php';
 
 // Include functions
 require_once PATH.INC.'/functions.php';
+
+// Start the session
+session_start();
+
+// Create a Login object
+$rs_login = new Login;
 ?>
 <!DOCTYPE html>
 <html>
@@ -23,11 +29,12 @@ require_once PATH.INC.'/functions.php';
 	<body class="login">
 		<div class="wrapper">
 			<h1><a href="/"><?php getSetting('site_title'); ?></a></h1>
-			<form action="" method="post">
-				<p><label for="username_email">Username or Email<br><input type="text" name="username_email"></label></p>
+			<?php $error = isset($_POST['submit']) ? $rs_login->userLogin($_POST) : ''; echo $error; ?>
+			<form class="data-form" action="" method="post">
+				<p><label for="username_email">Username or Email<br><input type="text" name="username_email" autofocus></label></p>
 				<p><label for="password">Password<br><input type="password" name="password"></label></p>
 				<p><label for="captcha">Captcha<br><input type="text" name="captcha"><img id="captcha" src="<?php echo INC.'/captcha.php'; ?>"></label></p>
-				<p><label for="remember_login"><input type="checkbox" name="remember_login" value="0"> <small>Keep me logged in</small></label></p>
+				<p><label class="checkbox-label"><input type="checkbox" name="remember_login" value="0"> <span>Keep me logged in</span></label></p>
 				<input type="submit" class="button" name="submit" value="Log In">
 			</form>
 		</div>